requirements.txt
streamlit
tensorflow
pillow
plotly
numpy